[NTDLL] Use LDR_LOCK_LOADER_LOCK_FLAG_RAISE_ON_ERRORS flag instead of TRUE [RTL] Fix RtlAddAce (the version checks were inverted.) [NTOSKRNL] Make SystemExtendServiceTableInformation case of NtSetSystemInformation behave more like Windows [NTOSKRNL] Fix PsGetCurrentThreadWin32ThreadAndEnterCriticalRegion
svn path=/trunk/; revision=60737

+++ reactos/lib/rtl/memstream.c
@@ -303,7 +303,7 @@ RtlCopyMemoryStreamTo(
     TotalSize = Length.QuadPart;
     while (TotalSize)
     {
-        Left = min(TotalSize, sizeof(Buffer));
+        Left = (ULONG)min(TotalSize, sizeof(Buffer));
 
         /* Read */
         Result = IStream_Read(This, Buffer, Left, &Amount);
